refactor(pt/politics-detail): modify file structures
