+++ packages/politics-tracker/components/politics-detail/default-text.tsx
