+++ packages/politics-tracker/components/politics-detail/toggle-lists/expert-point/index.tsx
@@ -1,6 +1,5 @@
 import styled from 'styled-components'
 
-import DefaultText from '~/components/politics-detail/shared/default-text'
 import ExpertItem from '~/components/politics-detail/toggle-lists/expert-point/expert-item'
 import type { ExpertPoint } from '~/types/politics-detail'
 
@@ -14,17 +13,13 @@ type ExpertPointProps = {
 export default function ExpertPoint({
   experts = [],
 }: ExpertPointProps): JSX.Element {
-  const expertItems = experts.map((value: ExpertPoint, index: number) => (
-    <ExpertItem expertItem={value} key={index} />
-  ))
-
   return (
     <Wrapper>
-      {!!experts.length ? (
-        <ul>{expertItems}</ul>
-      ) : (
-        <DefaultText title="專家看點" />
-      )}
+      <ul>
+        {experts.map((value: ExpertPoint, index: number) => (
+          <ExpertItem expertItem={value} key={index} />
+        ))}
+      </ul>
     </Wrapper>
   )
 }
